Move SNMP support code to separate module
